Not_usefull/RAG_Data_creation/requirements.txt
pandas>=2.0
requests>=2.31
wikipedia>=1.4
sentence_transformers>=2.5
faiss-cpu>=1.7.4 